deleteBook avec fs est unlink

+++ backend/controllers/book.js
@@ -1,5 +1,5 @@
 const Book = require('../models/Book');
-
+const fs = require('fs');
 
 // GET all books
 exports.getAllBooks = (req, res, next) => {
@@ -71,10 +71,23 @@ exports.postBook = (req, res, next) => {
 
 // DELETE a book
 exports.deleteBook = (req, res, next) => {
-    Book.deleteOne({ _id: req.params.id })
-        .then(book => res.status(200).json({ message: 'Deleted!' }))
-        .catch((error) => res.status(400).json({ error }));
-};
+    Book.findOne({ _id: req.params.id})
+        .then(book => {
+            if (book.userId != req.auth.userId) {
+                res.status(401).json({message: 'Not authorized'});
+            } else {
+                const filename = book.imageUrl.split('/images/')[1];
+                fs.unlink(`images/${filename}`, () => {
+                    Book.deleteOne({_id: req.params.id})
+                        .then(() => { res.status(200).json({message: 'Objet supprimé !'})})
+                        .catch(error => res.status(401).json({ error }));
+                });
+            }
+        })
+        .catch( error => {
+            res.status(500).json({ error });
+        });
+ };
 
 
 module.exports = router;
